django/requirements.txt
amqp==5.2.0
#asgiref==3.6.0
asgiref>=3.7.0
async-timeout==4.0.2
Babel==2.13.1
basicauth==1.0.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
#Django==4.1.5
Django==5.0.4
django-appconf==1.0.6
django-debug-toolbar==4.3.0
django-redis==5.4.0
django-tinymce==3.6.1
ghp-import==2.1.0
#gunicorn==20.1.0
gunicorn
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.5.1
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-material-extensions==1.3
mysqlclient==2.1.1
packaging==23.2
paginate==0.5.6
pathspec==0.11.2
platformdirs==4.0.0
prompt-toolkit==3.0.36
Pygments==2.16.1
pymdown-extensions==10.4
python-dateutil==2.8.2
python-gettext==5.0
pytz==2022.7.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
redis==4.4.2
regex==2023.10.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.3
urllib3==2.0.2
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
pyjwt